libepoxy: include egl in cflags for pkg-config based builds
Build systems that depend on pkg-config for cflags fail to find egl headers.
The example I ran into is building qemu in a dev shell environment. This
seems to be caused by nixpkgs patching the behavior of pkg-config so that it
does not include private dependencies in the cflags output unless it's a
static build. I also see examples of this fix in telepathy-glib, ubports-click,
libgweather, and cairo.